When resolving expression types, treat as nested functions only if the expression refers to functions / methods
Summary:
Currently, we treat an expression as a nested function, even if the expression
is just a local variable. See the test whose expected result is changed in this
diff.

Performance-wise, this is likely cheap (although calling
`Target.get_module_and_definition` could be expensive) because:
1. when calling `Target.get_module_and_definition` to load the source code, the
source code is already loaded because the nested function and the function that
is being analyzed when calling `resolve_ignoring_errors` are in the same
module.
2. this case is rare. See the log of D70103254, which counts the number of
times hitting this case when analyzing IG -- this is never hit!

Reviewed By: arthaud

Differential Revision: D70115440

fbshipit-source-id: 05d5ae72a1fb80d8097ab24eda1f202cafd9bc5a