Bind function parameters using Var
Summary:
Instead of deciding on the type of a function param during bindings, try to wait.

We set up function bindings like lambda bindings, using vars if there is no annotation. Then, in the answers phase, we try to figure it out with more information.

For now, I am not using any additional information. We pass in the self type and use that for the first parameter, if it is unknown.

From here, my plan is to use the decorator context (classmethod, staticmethod) to determine if the type of the first parameter should be `Self` or `type[Self]`.

There is one wrinkle here. Unlike LambdaParameter, where we will always solve the entire lambda at once, we might observe FunctionParameter bindings before we solve the function binding. To deal with this, I add the function binding's index to the parameter's binding, and force the function to be solved before the parameter. This is enough to ensure that Var-style function parameters are always solved before use, so the Var never leaks out.

Reviewed By: stroxler

Differential Revision: D70050192

fbshipit-source-id: d640f584341fb45e0cd041b1bac6a5a1d4b5308b