Move the TypedDict check up and put a TODO in
Summary:
The typed dict class initialization check does not depend on `value_ty`, which
is important to highlight because the series of mutations of `value_ty` is complicated
and is tied in with some buggy (or in some cases, like whether enum values are literals
or promoted types, not exactly buggy but maybe accidental) behavior.

While I'm at it, put in a TODO; as per a conversation with rchen152 a few weeks ago,
we don't think it's actually well specified whether a TypedDict is permitted to contain
class-level attributes like class and static methods. Semantically it makes perfect sense
to allow this; TypedDict *instances* don't really exist as class instances, but the
class object itself is a fairly normal Python class object.

The TODO will remind me to post this in the typing discourse when I have more
time.

Reviewed By: grievejia

Differential Revision: D70256959

fbshipit-source-id: e24de2baee796ed37fe390ae3098fb30fc200410