Fix a bug where we failed to promote literal attr types
Summary:
We have a complicated performance hack in the `ClassField` code, where we try
to skip promoting literals if an annotation is present that we will eventually
use to overwrite `value_ty`.

But the performance hack is incorrect - if the annotation doesn't contain a type,
only qualifiers, then we will in fact continue to use the original type
which means that the literal leaks out into attribute type inference.

This is both incorrect and a conformance failure. Fix it by making the perf
hack do a more complex check for non-None type *inside* the annotation.

In the process, we are for the moment losing literal types on `Final` attributes,
because we were relying on the buggy logic removed here to prevent promotion.
I marked a test as a bug with a TODO to fix this.

In spite of the new bug I am sure this is better - failing to preserve a
literal is a much less severe problem than incorrectly preserving one (literals
are a fairly recent and advanced type system feature, far less code is likely
to be impacted by failing to preserve them in every code). I'll try to fix
the regression later as I refactor ClassField.

Reviewed By: grievejia

Differential Revision: D70256144

fbshipit-source-id: 58db1a6b8c23c1a6fbf26272a97bb0c66fd5016a