Track range of scopes for autocomplete purposes
Summary:
For autocompletion purposes, we need to answer the question of:

> At this position, what are the set of available definitions we can use

To answer the question, this diff sets up an optional trace (under `enable_trace`) that will retain a tree of scopes that track their ranges. With this information, we can do a range-limited tree traversal to find the set of all definitions.

In the next diff, the function will be used to power variable autocomplete.

Reviewed By: stroxler, ndmitchell

Differential Revision: D70282227

fbshipit-source-id: 5a62bdbf3afef4b4a03ee44da4cb4c18011fd6c7