Add a test case for downstream use of annotated member
Summary:
There was no test that made it clear how exactly we handle downstream use
of annotated enum members.

Our handling is actually incorrect. In this code:
```
import enum

class E(enum.Enum):
    X: int = 1

print(E.X)
print(E.X.value)
```
it turns out that `E.X` is indeed an enum member, the annotation does not
prevent the enum magic from happening.

The code is not valid, but it could easily happen in third party libraries.
This is not a minor consideration - Pyre1 used to handle this wrong and we
had to modify many examples (including some in Pyre1's python frontend itself
if I remember correctly!) to fix the behavior, so we should *expect* to
encounter third-party libraries with this behavior and handle them
correctly with respect to the runtime.

I found this bug because I want to make refactors tied to the behavior, which
means I need the behavior to be well-defined on all edge cases.

Reviewed By: rchen152

Differential Revision: D70326316

fbshipit-source-id: 136235625654aad5e670275e314b68566e3d10c8