Correctly infer that an annotated enum member is a member
Summary:
This diff moves the annotation-related transform of the class field type to happen
*before* the enum-related transform, which has the impact of fixing a bug where we
were dropping the enum member-ness of annotated members (because we were overwriting
the synthesized enum literal with the annotation).

The resulting behavior - which still throws a type error complaining that you aren't
supposed to annotate enum members - now correctly analyzes such members, which is
much better behavior given that we should expect this issue in third-party code.

I also added the `.value` attribute to the test, because it's good for us to have coverage
of the fact that the annotation (in this example `float`) takes precedence over the
inferred expression type (in this example, `Literal[5]` which would be promoted to `int`).

Reviewed By: rchen152

Differential Revision: D70326330

fbshipit-source-id: 65435f961a10ee94e1f273afe5e87fe5fc1d2701