fix crash in type argument defaults when there's a mismatch
Summary:
The bug is that the number of defaults isn't based on how many parameters have been processed, and the last branch doesn't know when previous branches have inserted defaults.

This diff moves the TypeVarTuple and ParamSpec default handling to the last branch, inserting the appropriate kind and number of defaults based on the remaining parameters.

Note that if there's only one TypeVarTuple left at the end, it can be empty so we shouldn't emit an error.

Reviewed By: rchen152

Differential Revision: D70358494

fbshipit-source-id: 60a720ddc39409312c3cf7e90a458a2dee278233