Add a `classproperty` unit test, which partially passes
Summary:
Since the user-defined descriptor that matters most in a couple of codebases
we are targeting is a class property, we ought to cover that case with a
unit test. It requires that our handling deals with generics and can properly
propagate them from the decorator application to the method call. This all works
out-of-the-box but it's good to have that explicitly demonstrated in our unit tests.

The unit test partially passes: the descriptor binding behaves as expected
(we evaluate the `__get__` call on reads, and complain on writes), but as
expected the actual decorator application fails because we have not yet
implemented a mechanism to correctly infer the parameter type of the
`cls` parameter in the function definition: we automatically infer `Self`
for the first parameter of a function def in a class body by default, which
does not agree with the restriction on `classproperty.__init__`.

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D70500787

fbshipit-source-id: 81370ba881ab908a09f9668cdfb4c680efbbe645