Register the defines and call graphs of decorated targets into shared memory
Summary:
The benefits are:
1. Avoiding passing `CallGraph.DecoratorResolution.Results` to each worker in
higher order call graph fixpoint, which can be a huge map and hence cause huge
performance problems.
2. Better readability, since the querying of defines and call graphs for
decorated targets can now be done in the same as normal targets.

This diff also moves the registering of the call graphs (of decorated targets)
outside of `build_whole_program_call_graph`, so that we explicitly define:
1. registering defines into `Target.DefinesSharedMemory`
2. registering call graphs into `CallGraph.SharedMemory`

The only downside is to move `DecoratorResolution` after defining
`CallGraph.SharedMemory`. It would have been nice to keep
`CallGraph.SharedMemory` as the last module in `callGraph.ml`, since this is
the most important module.

Reviewed By: arthaud

Differential Revision: D70370192

fbshipit-source-id: 82ddcd66ca72873275eb0e711da049ff56a61b37