Keep track of the rdeps
Summary:
Given a large example, it took ~300ms to do an rdeps walk. That's too slow for our goals, so we must cache the rdeps.

We could optimise this for one-shot, and not even store the rdeps. But I can't really measure any delta in time taken on a large example (within the margin of error), so keeping it simpler for now.

Note that as of now, simply having the rdeps doesn't make things massively faster. We need some more tricks.

Reviewed By: grievejia

Differential Revision: D70505598

fbshipit-source-id: 3a80943951d728b1b7740b2d912076c1f55067bc