Fix tp local variable issue
Summary:
The issue comes from here https://www.internalfb.com/intern/qa/783048/pyre-does-not-recognise-class-type-parameters-used

The reason we get an error is that we did not remove the class type parameters from the unbounded variables of a function. (we only remove function type parameters from unbounded variables of a function).

The easiest place I found to fix this was in typecheck although open to any alternative thoughts.

Reviewed By: grievejia

Differential Revision: D70584274

fbshipit-source-id: 5d11530b3f22426ab457192a188ecfde1e8f1d1f