Fix bug in `is_private_field`
Summary:
Dunder attributes are not private.

The main manifestation of the fix is that variance inference changes
which is a very minor change given how little variance inference is used,
but the bug came up when investigating a user reporting a case where
we are failing to check for override consistency on `__init__`; fixing
the bug here helps make it more obvious that this function is not the
root cause of the reported issue.

Reviewed By: migeed-z

Differential Revision: D70532281

fbshipit-source-id: 4c8c19ee3c8d950d29e41a3f9797a5fca573ae10