Add a test case of a bug in is_private_field
Summary:
Jia and I mistakenly identified `is_private` as a likely root cause
of a bug a user is experiencing where Pyre fails to check `__iter__`
for override consistency.

It turns out that the bug is *only* in `is_private_field`, not in
the (newly renamed) `is_mangled_private_field`, and as a result the
bug we saw does not seem to be the root cause of this user's problems.

Nontheless, we ought to fix the bug: `is_private_field` is mistakenly
identifying double-underscore fields as a private. This primarily manifests
in incorrect variance inference.

This diff extends our existing test case for private attributes to show
the bug, I'll fix it in the next diff.

Reviewed By: migeed-z

Differential Revision: D70532280

fbshipit-source-id: 17a28972dd089818e942bea421858c7a45c32615