Make hover work on definitions
Summary: Currently hover doesn't work on definition site, which feels very inconsistent. In this diff, I created some helper based on the existing function to get all available definitions at a position, to tell whether a given location is a definition. If it is, then we can directly use the key to get the type for hover.

Reviewed By: ndmitchell

Differential Revision: D70642660

fbshipit-source-id: 15b3cab9b885c644a32b7453512b6ccda22a8856