Make go-to-definition on definition jump to itself
Summary: Based on the infra created in the previous diff, we can now make go-to-definition on definition jump to itself. This is a more helpful behavior in general, since it ensures that we don't return null results, and one day when we implement find-references, vscode will then jump to references locations instead in this case.

Reviewed By: ndmitchell

Differential Revision: D70643227

fbshipit-source-id: 32eae996739b93f0962c8958e3ab7d0b5291142c