Stop propagating sources and sinks when reaching the maximum_source/sink_distance from a rule
Summary:
# Context
We support specifying a `maximum_source_distance` and `maximum_sink_distance` on rules. This simply discards issues that have the source or sink with a distance larger than the specified value.
However, we still propagate those sources and sinks during the taint analysis, leading to performance issues.

# This diff
This diff disables the propagation of sources and sinks that reached the `maximum_source_distance` or `maximum_sink_distance` specified by rules. This only applies if the source/sink is only involved in rules with a `maximum_source_distance` or `maximum_sink_distance` set.

Reviewed By: tianhan0

Differential Revision: D70629513

fbshipit-source-id: 5ff6edb2c5df6b06a4e1b390d06f0a625a146c32