Optimise ErrorKind string conversion
Summary: This used to take 10s in a benchmark, where the total benchmark time was 25s. Ouch. Now it doesn't show up on the profile at all and is 10s faster. The Kebab conversion is very robust, dealing with UTF8, and not very well optimsed. We can optimise it, but actually we'd rather not have to do an allocation at all, so might as well cache it and then we can `&'static str` it.

Reviewed By: yangdanny97

Differential Revision: D70699279

fbshipit-source-id: 12041a0de440d14706841007eb09bb1265fcd744