Use ErrorStyle::Delayed for filesystem loaded module and limit published diagnostics to opened files for now
Summary:
Currently, we have a bug where we see no errors after we open a file that's checked initially but not opened initially.

It happens because the module is forever in the state of having filesystem path, and for filesystem path we will check it with `ErrorStyle::Never`. This is reasonable if we want to limit the published diagnostics to opened files (it's even necessary for now to not freeze VSCode since we still generate a tons of errors due to type checker bugs and incomplete support). However, the issue is that when we opened the file, we never properly promote the module to have `ErrorStyle::Delayed`.

This diff workarounds the problem for now by making these fs-loaded files have `ErrorStyle::Delayed`. To avoid pushing tons of errors to IDEs for now, I also limited the diagnostics we publish to opened files. The limit is basically a less buggy version of the status quo.

Reviewed By: ndmitchell

Differential Revision: D70710534

fbshipit-source-id: f855e2543aa0a0a51893df0667bebcdf36bf1be0