Remove flow errors on pyre2_wasm.js
Summary: We currently have 0 flow errors before the wasm build. However, after building the wasm file and copying it over to the website folder, there's a large amount of errors from `pyre2_wasm.js`. This diff removes them by adding the file into `.flowconfig`. Given that for development, we will be expected to build the wasm first and then make changes to the website, it's much easier if we can exclude all the flow errors from the wasm file

Reviewed By: stroxler

Differential Revision: D70717195

fbshipit-source-id: 0618e70a2f43218244742800b7d2a65864051b64