Update formatting for sandbox error message
Summary: This diff updates the formatting for error message to match how we output it in cli. Added a TODO in code to simply return the formatted string from binary, so we don't need this duplicated logic. However, I think it's worth landing this logic for the purpose of the demo tomorrow.

Reviewed By: rchen152

Differential Revision: D70748557

fbshipit-source-id: ed681f9c417de8655f4b95ede2ef430b5f60bb99