Make internal docs internal only in docs sidebars
Summary:
The github deploy script is broken after D70738995. Example: https://github.com/facebook/pyrefly/actions/runs/13712567973/job/38351694383

This seem to be because we are trying to reference an internal fb doc in `sidebar.js`, but folders with `fb` does not get sync'd to github. The right way to do this would be to wrap `fbInternalOnly` around internal only docs in `sidebar.js`

Reviewed By: stroxler, yangdanny97

Differential Revision: D70760788

fbshipit-source-id: 2b539cc52cb28043467e0a34bbe4c9a1e046e0df