Print hash table stats when `Hash_table_full`
Summary:
Currently, we just see `Hack_heap.SharedMemory.Hash_table_full`, which is not
informative about what is going on under the hood in `hh_shared.c`. This diff
enables printing more stats about the hash table itself.

The change in `worker.ml` seems necessary in order for the `fprintf` outputs to
be printed. I think this is because the master process needs to get the
backtrace of the workers.

Reviewed By: arthaud

Differential Revision: D70733939

fbshipit-source-id: 9404dfc3a1123868a3a809e1b112e0e906a802b0