Include type as optional detail for dot autocomplete
Summary:
Continue from the work in D70731782, this diff adds in type information for dot autocomplete. We do it as a post-processing step after we have grabbed the items, and call relevant function that turns an attribute into a type for every item.

Since it can be expensive, this is only enabled for autocomplete. Go-to-definition for attributes that goes through the same path won't try to get types, since they are not used anyways.

Reviewed By: ndmitchell

Differential Revision: D70741057

fbshipit-source-id: b633d4afa9a1eb284f79812b90a65ed5a67c7393