Allow the playground to be used for different files
Summary: The current playground assumes the existence of a single model, so it breaks down badly when the docs page introduces a lot of editors. This diff removes the assumption by maintaining a map from model to IDE provider functions, so that every one of them can work independently. To correctly find the model in TryPyre2, we will require the component be given a unique path.

Reviewed By: stroxler, dluo

Differential Revision: D70800628

fbshipit-source-id: 0eef458559b49a12afffd346148aea395e152c2d