Remove unneeded function from public API
Summary:
In order to implmement caching of the types query, I need to give
`Query.ml` easy access to the *reference* of a module at a path
(because that's how invalidation keys work, which means the cache
will need to be based on reference).

Making the changes to allow this requires doing a little bit of tetris
on the function interfaces, and having an unnecessary helper function
in the public API gets in the way. Let's just strip it.

Reviewed By: yangdanny97

Differential Revision: D70797517

fbshipit-source-id: 4ee88f6305e50a5c1b54980d96bdcc9fc1a81de5