Add the ability to fetch invalidated type checks from an UpdateResult
Summary:
We have a static analysis partner team that wants the `types` query - which these days has
to re-check modules, because due to limitations in Pyre architecture that's the only good
way to correctly handle expression scope - to be fast if it is run twice on the same module
(which is a common thing to do when iterating on the downstream static analysis tool: make
a change and reload but with the same Pyre server, and re-fectch types).

This diff moves us toward being able to have a cache in the Pyre server for types queries,
by making it possible for an UpdateResult to tell us when the cache was invalidated.

Because everything is keyed on qualified name, we have very poor ability to actually
figure out what is a module name. As a result, the functionality grossly over-approximates
potentially changed module names by just taking all parent references of every changed
function reference. This is ugly, but works for our needs; I added a detailed comment
block making this clear.

Reviewed By: grievejia

Differential Revision: D70797510

fbshipit-source-id: 720af5cdb120d57df9469a4d4c3241ce7e8a9999