Expose the qualifier logic to Query.ml
Summary:
The previous diff refactored `LocationBasedLookup` functions
so that the `*_for_path` functionality was implemented in
terms of a bind over looking up the path plus a `_for_qualifier`
simpler function.

This diff prepares to wire a query cache into `Query.ml` by
making it actually do the bind operation directly, so that it
has access to the `qualifier` that will be our cache key.

Reviewed By: grievejia

Differential Revision: D70797518

fbshipit-source-id: b07681b9800217772afdd71743e0488c9dc0e9d7