Make `find_all_resolved_types` qualifier-keyed
Summary:
We invalidate type information based on `Reference`, not on path,
which means that for our query cache we want to key it by reference
(that is, the module name `qualifier` in Pyre1 terminology).

This means we want to expose the qualifier to the `Query` module,
so we want the `locationBasedLookupProcessor` interface to be
based on qualifier rather than operating directly on raw paths.

To do this, I teased apart the API so that the bind over looking
up a qualifier is a helper function; in the case when we need to
cache we'll be able to use the underlying functionality directly.

Reviewed By: grievejia

Differential Revision: D70797522

fbshipit-source-id: e5e35023949e0696b66ced7d07d540b07cbeaead