Implement cache wiping
Summary:
I wanted to get this in place before we actually use the cache, so that there's
never a known buggy state.

Because we can only produce type errors for modules that have type check results,
tracking invalidated type check keys should be sufficient.

Note that if a module is deleted, it remains part of our dependency graph (the
Pyre dependency graph is grow-only, it never shrinks) which means that there should
be no bug in cases where, for example, a module is deleted and then re-added: that
module name will show up in `modules_with_invalidated_type_check` in both updates,
so the cache for that module will be empty after both updates.

Reviewed By: grievejia

Differential Revision: D70797516

fbshipit-source-id: 81d41c905acf318167d0d118551cae839f69bd84