Wire up up cache invalidation plumbing
Summary:
At this stage our query cache is still just a dummy value, but we create
a public API to invalidate the cache based on an environment update result,
and the `requestHandler.ml` code makes the necessary call.

Note that the server code assumes permanent components to the state throughout,
so both populating and invalidating the cache will have to be mutable operations;
this was already true but is made more obvious by the shape of the API we create
in this diff.

This should be the last plumbing diff - at this point we can begin implementing
and using the cache in `Query.ml`, and all the logic should be isolated there
behind clear abstraction boundaries.

Reviewed By: grievejia

Differential Revision: D70797511

fbshipit-source-id: 6567db2859f0ef0298f6c6c0a4b901844fab993b