Bootstrap a query cache type
Summary:
It doesn't do anything yet, but I want to plumb the type through
the server code before implementing any nontrivial functionality.

The type is abstract in the signature, my intention is to hide all details
inside the `Query` module.

I'm slicing diffs pretty thinly here for ease of review, because (a) the
cache logic I am aiming to add here will be impractical to unit test and
(b) I'm a bit rusty at working on the server code.

Reviewed By: grievejia

Differential Revision: D70797514

fbshipit-source-id: 49be37355a9c3f879e972d5b307a7dad6ba5f70c