Fix up sandbox monaco model load logic
Summary:
Current sandbox load is hard to read and buggy. There are a couple issues:
1. if you open the sandbox, navigate away and come back, there's no backing monaco editor model and IDE completely breaks
2. if you load the sandbox with an initial error, it would not show errors until an initial keystroke

This diff cleans this up by cleaning up the react lifecycle. We fetch the current model in 2 places: - - `onMount` (handle when user leaves the screen and comes back)
- `onDidCreateModel` for the initial load, since often times, the monaco model was not created at the time of onMount

Also moved the logic for force updating the model outside of useEffect, since we only care about calling it when we first open the page.

With these changes, we are now safe to add `reveal_type` back into the example

Reviewed By: yangdanny97

Differential Revision: D70803277

fbshipit-source-id: 324d1eab760c212e14fc68cd497515a745c62445