Avoid unnecessary Box on NarrowVal
Summary:
Given that all NarrowVal's had a Box around them, might as well eliminate that, and store the value inline. Avoids one additional malloc.

That makes NarrowOp bigger. By adding a Box in two places we limit that increase to 1 word, which seems fine. In order to not bloat Binding we now put NarrowOp into a Box.

Reviewed By: yangdanny97

Differential Revision: D70886559

fbshipit-source-id: f44c4ae537d84f9a7b2c6fdfe7a54acc13f00f36