handle generic TypeGuard
Summary:
We can't just look at the return type of the function when doing narrowing for type guards, because TypeGuard can be parameterized with a type variable, and the narrowed type can depend on one of the parameters.

This means we should actually check the call, and unwrap the return type.

This diff increases the size of bindings - should I box the type, or does it not matter?

Reviewed By: ndmitchell

Differential Revision: D70813134

fbshipit-source-id: f737e458031b9d7c6d404558a2f80e76139a020f