Add an end-to-end test of query cache invalidation
Summary:
The scariest possibility with our new query cache is that cache
invalidation does not work.

Here, I add a simple unit test of an end-to-end running server
where we update the code and observe that the `types` query output
correctly changes for both the directly modified file `a.py` and
a dependent module `b.py`.

Reviewed By: grievejia

Differential Revision: D70797512

fbshipit-source-id: eb3c9b576019d063ea649d14b96b9227301643e9