Wire in Configuration parsing
Summary:
Attempt parsing a provided config in project checking mode, failing if it can't be parsed.

For project checking mode, pass in the files included in `ConfigFile.project_includes` as the files to check.

NOTE: until the next diff, all paths are relative to the CWD, not the directory containing the config.

Reviewed By: Solumin

Differential Revision: D70898540

fbshipit-source-id: f5992c795e8833f5c7f117a0f58e3b869c67b39d