example of type form validation
Summary:
There's a ton of these rules, I can either implement them here or as methods on the TypeFormContext enum (or move them somewhere else after it's all done).

It's probably too complex to handle as a simple switch without duplicating code. I'll need to take a look at the logic for untype  to see if things need to be rearranged.

Reviewed By: rchen152

Differential Revision: D70941692

fbshipit-source-id: ad4128f9b145334d1f7dcb51f5aef92f04b9be51