Make sandbox prettier on small and large screens by implementing dynamic height
Summary:
The sandbox height is still hardcoded to 600px for the code section currently. This diff makes the height dynamic according to screen size. It's still not perfect, and is slightly buggy when there's screen resizes, but this is already a large improvement to the current state.

I attempted to use `automaticLayout: true` for the monaco editor and not pass in a hardcoded height, but it seems like this causes the height to default to 5 and completely hides the editor.
I also took a look at some of the existing posts for monaco-editor on github, but none of the solutions seem to have worked so far. Nevertheless, I believe it's better to make this intermediary fix here since the actual solution will take longer to implement.

Reviewed By: MaggieMoss

Differential Revision: D70911043

fbshipit-source-id: 33a05cdb2b52db70a75092d44c47e8a58cb43491