Support attrs defined in unit test setup methods (copy Pyre1 logic)
Summary:
As I was looking at the results of running Pyrefly on a real application, I noticed that
a common error is attribute not found, and after digging into a random sample of 3 examples,
*all three of them* were attributes defined in a test setup method.

This led me to check how Pyre1 handles this, and it uses an allowlist of test setup methods
that get treated "like constructors" in that they are allowed to define instance attributes
via assignment to `self`.

We should do the same in Pyrefly. I actually think that heuristics based on naming conventions
are the right way to handle this, rather than trying to use some kind of type-based analysis
that specifically handles cases like `unittest.TestCase`, because (a) naming conventions for
testing are very strong conventions and (b) we really don't want our solution to be overly
tied to a specific testing framework. It also happens to be much easier to implement a naming
convention based solution anyway, since we need to handle this in `bindings` :)

Reviewed By: grievejia

Differential Revision: D70978112

fbshipit-source-id: 247762ac351c2d01f0cb46ac50ce95b50015291d