Simplify `class_key` creation
Summary:
It's a little strange to have `class_name` floating around in the bindings code
and have it be a `ShortIdentifier` rather than a `Name`. In addiiton, it's kind of
silly to be making it and then using it exactly once, to create the `class_key`, with
a `clone()`... we can just inline this, saving both the clone and the extra variable
floating around the namespace of fairly complicated functions.

Context: This is a nice small readability win on its own, but it came up
because I got confused about what `class_name` was when looking at a tweak to
attribute binding so that we can better support Python unit testing patterns.

Reviewed By: rchen152

Differential Revision: D70978113

fbshipit-source-id: 70bb5cf5ff893139b0ba7a8e0c14d368fbc64c66