Never load from memory from `LspLoader.find`
Summary:
`Loader.find` is for module resolution purposes: given a module name (dot separated identifiers), find the actual module. We don't want in-memory files to ever be resolved in this way, because then other modules could depend on in-memory files and cause expensive recheck on every keystroke.

With this change, how does the system know the up-to-date in-memory files? Well, in memory files are never discovered through `Loader.find`, but explicitly given to `State::run`, at the time we already know both the module name and the module path (always `ModulePathDetails::Memory(_)`, so it knows that it needs to read content from memory).

Reviewed By: ndmitchell

Differential Revision: D70977744

fbshipit-source-id: beb04a11fe6846c5117dc5bf788867c81d38978e