Fix Pyrefly internal error when importing non existing module
Summary:
## Problem
https://github.com/facebook/pyrefly/issues/18
Currently, whenever an invalid import statement is entered in the sandbox, the sandbox crashes and fails to do anything else in the future. This is because we have a line of `panic` in `playground.rs`

## Solution
Update `Panic` with a `FindError`. Also add tests to ensure that this behavior is tested.

Reviewed By: ndmitchell, MaggieMoss

Differential Revision: D70984463

fbshipit-source-id: 7c63480bf78a45561dd6e3e24f11a2cad20440e8