Make `CheckArgs::run_watch` async
Summary:
I'm trying to implement file watcher with `watchman` (see D70984788), but one issue I ran into was that the underlying watchman APIs require to be invoked as async code, whereas the current file watcher API is sync.

This diff, along with D70984791, proposes we turn the pre-existing file watcher API into async functions, as a way to bridge the gap. This requires that in OSS check command watch mode, we need to create an `async` entry point by firing up a tokio event loop.

Reviewed By: ndmitchell

Differential Revision: D70984787

fbshipit-source-id: 4990d86ff9eb1317f211309952ee3514aadb9fce