Make `Watcher::wait()` async
Summary:
I'm trying to implement file watcher with `watchman` (see D70984788), but one issue I ran into was that the underlying watchman APIs require to be invoked as async code, whereas the current file watcher API is sync.

This diff, along with D70984787, proposes we turn the pre-existing file watcher API into async functions, as a way to bridge the gap. This requires to turn the `Watcher::wait()` function into async, and adjust the callers slightly to await the result.

Reviewed By: ndmitchell

Differential Revision: D70984791

fbshipit-source-id: a189d07fdb13d9428d0feefec5b073b23be19e66