Rename project_include config item to project_includes
Summary:
This diff renames `project_include` to `project_includes`, since I think it sounds better and makes more sense here as a name, given we're working with a list of globs.

We could keep CLI args singular and ENV/config items that are lists plural, but I like the idea of keeping it consistent everywhere. Please let me know if you disagree.

Reviewed By: stroxler

Differential Revision: D71146503

fbshipit-source-id: 6f1ba3a122b8147062a4949365e8a33d58d066b0