Implement --summarize-errors
Summary:
This is modeled after wut.py, an internal script for summarizing the errors found in a directory.
It displays the top 30 directories and their errors, the top 30 files and their errors, and the top errors, all by error count.

While the funcitonality is all there, the formatting could use some polish.

Reviewed By: ndmitchell

Differential Revision: D71158903

fbshipit-source-id: 2cb3f9097971cd1fb746260a7f797fdeb999f13e