Make a small test case to repro a bug
Summary:
An an example real codebase, we encountered some decorators that are the result
of constructors acting as decorator factories and producing callable instances
of the decorator.

For reasons that are not yet clear, Pyrefly is failing to understand that these
instances are callable (we have at least some support for `__call__`, but somehow
the lookup is failing).

This test reproduces the bug in a very small example that we can debug

Reviewed By: ndmitchell

Differential Revision: D71290321

fbshipit-source-id: cc21e798a1f10ab77871925136b2b50ed24b1451