Add constructor with glob from root
Summary:
The next couple of diffs will be setting up configurations so that they'll correctly map out files when run from a directory other than where the config is defined. We can do this by changing all relative paths in the config to prepend the path to the config, and appending the path to the config to the end of `search_roots`.

This diff sets up a helper constructor that rewrites a `Globs` to prefix a certain root onto the given patterns, using the `new_with_root()` constructor, but without copying the patterns contained in the globs. This will be useful in upcoming diffs when we're parsing a config from a  path and need to rewrite all globs to be read from the given path.

Reviewed By: stroxler

Differential Revision: D71213041

fbshipit-source-id: 14372de9b155c7fc4ca6a5001963dcbb0007ca39